fixed quest code formatting, quests can now be impossible
whenever the quest book opens, it will now check which quests are possible currently and remove them from the display. quests that are not possible to complete within the current world will not show in the book and will not display in the chat if they are unlocked (they will still be unlocked, however. just in case something is cheated in or changed in their world to allow it to be completed).